Fix LineType assumption problem
Before this change, when trying to locate the account to credit and
the total amount to allocate, the script assumed that the LineType
from APInvoiceGLDetails would be 'DI.' This was an incorrect
assumption because line items on invoices generated as the result of a
PO have LineType 'PO.' An examination of the database indicates that,
for now at least, the possible line types are 'AP' (supplied by
default when the invoice is created; read-only), 'DI' (indicates the
invoice originated in AP--manual entry?), and 'PO' (indicates invoice
originated in PO).

After this change, the script ignores 'AP' LineTypes and uses the last
detail line item's account and amount information. Essentially, this
implementation assumes that there is only one pertinent detail line.